Refacto generation coordinates on free cells
